+++ src/textblob/en/taggers.py
@@ -3,7 +3,6 @@
 import nltk
 
 import textblob as tb
-import textblob.compat
 from textblob.base import BaseTagger
 from textblob.decorators import requires_nltk_corpus
 from textblob.en import tag as pattern_tag
@@ -17,7 +16,7 @@ class PatternTagger(BaseTagger):
 
     def tag(self, text, tokenize=True):
         """Tag a string or BaseBlob."""
-        if not isinstance(text, textblob.compat.text_type):
+        if not isinstance(text, str):
             text = text.raw
         return pattern_tag(text, tokenize)
 
@@ -30,7 +29,7 @@ class NLTKTagger(BaseTagger):
     @requires_nltk_corpus
     def tag(self, text):
         """Tag a string or BaseBlob."""
-        if isinstance(text, textblob.compat.text_type):
+        if isinstance(text, str):
             text = tb.TextBlob(text)
 
         return nltk.tag.pos_tag(text.tokens)