+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Removals:
 
 - ``TextBlob.translate()`` and ``TextBlob.detect_language``, and ``textblob.translate`` 
   are removed. Use the official Google Translate API instead (:issue:`215`).
+- Remove ``textblob.compat``.
 
 Support: